Refactor emle.models.
